UPDATE: D_arena creation
We keep all battles between models in the tiers and also retain battles between models of the same capacity bin, because we otherwise end up with just 19K samples. The dataset the authors used does not seem to exist. This is a first attempt at a workaround or understanding what they actually did.